server/pyproject.toml
[project]
name = "finsight-server"
version = "0.1.0"
description = ""
authors = [
    {name = "shaik asif",email = "smd.20sa@gmail.com"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.121.3,<0.122.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "supabase (>=2.24.0,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "langchain (>=1.0.8,<2.0.0)",
    "langchain-openai (>=1.0.3,<2.0.0)",
    "langgraph (>=1.0.3,<2.0.0)",
    "tiktoken (>=0.12.0,<0.13.0)",
    "pypdf (>=6.4.0,<7.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-core (>=1.1.0,<2.0.0)",
    "cohere (==5.11.4)",
    "pdfplumber (>=0.11.8,<0.12.0)",
    "tavily-python (>=0.7.15,<0.8.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
